slide-120

+++ backend/routes/posts.js
@@ -38,7 +38,8 @@ router.post(
     const post = new Post({
       title: req.body.title,
       content: req.body.content,
-      imagePath: url + "/images/" + req.file.filename
+      imagePath: url + "/images/" + req.file.filename,
+      creator: req.userData.userId
     })
     post.save().then( createdPost => {
       res.status(201).json({
@@ -104,11 +105,18 @@ router.put(
       _id: req.body.id,
       title: req.body.title,
       content: req.body.content,
-      imagePath: imagePath
+      imagePath: imagePath,
+      creator: req.userData.userId
     });
-    console.log(post);
-    Post.updateOne({_id:req.params.id}, post).then( result => {
-      res.status(200).json({message: "Updated successfully!"});
+    //console.log(post);
+    Post.updateOne({_id:req.params.id, creator: req.userData.userId}, post).then( result => {
+      if( result.n > 0 ) {
+        res.status(200).json({message: "Updated successfully!"});
+      }
+      else{
+        res.status(401).json({message: "Unathorized!"});
+      }
+
     });
 });
 
@@ -116,11 +124,16 @@ router.delete(
   '/:id',
   checkAuth,
   (req, res, next) => {
-    Post.deleteOne({ _id: req.params.id }).then(result => {
-      console.log(result);
+    Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+      if( result.n > 0 ) {
+        res.status(200).json({message: "Deleted successfully!"});
+      }
+      else{
+        res.status(401).json({message: "Unathorized!"});
+      }
     });
-    console.log(req.params.id);
-    res.status(200).json({message: "Post deleted!"});
+    //console.log(req.params.id);
+    //res.status(200).json({message: "Post deleted!"});
 })
 
 module.exports= router;
